Split record cleanup unit tests Add record cleanup unit test helper
