+++ test/test_rosbag/test/record_sigterm_cleanup.py
@@ -38,61 +38,23 @@ import os
 import unittest
 import rostest
 import sys
-import time
 import signal
-import subprocess
+from record_signal_cleanup_helper import test_signal_cleanup
 
-RECORD_COMMAND = ['rosbag',
-                  'record',
-                  'chatter',
-                  '-O',
-                  '--duration=5']
-SLEEP_TIME_SEC = 10
+TEST_BAG_FILE_NAME = '/tmp/record_sigterm_cleanup_test.bag'
 
-class RecordCleanup(unittest.TestCase):
-
-  def test_sigint_cleanup(self):
-    """
-    Test that rosbag cleans up after handling SIGINT
-    """
-    test_bag_file_name = '/tmp/record_sigint_cleanup_test.bag'
-    test_signal_cleanup(test_bag_file_name, signal.SIGINT)
-
-    # check that the recorded file is no longer active
-    self.assertTrue(os.path.isfile(test_bag_file_name))
-    self.assertFalse(os.path.isfile(test_bag_file_name+ '.active'))
+class RecordSigtermCleanup(unittest.TestCase):
 
   def test_sigterm_cleanup(self):
     """
     Test that rosbag cleans up after handling SIGTERM
     """
-    test_bag_file_name = '/tmp/record_sigterm_cleanup_test.bag'
-    test_signal_cleanup(test_bag_file_name, signal.SIGTERM)
+    test_signal_cleanup(TEST_BAG_FILE_NAME, signal.SIGTERM)
 
     # check that the recorded file is no longer active
-    self.assertTrue(os.path.isfile(test_bag_file_name))
-    self.assertFalse(os.path.isfile(test_bag_file_name+ '.active'))
-
-
-def test_signal_cleanup(test_bag_file_name, signal_to_test):
-  """
-  Helper method to start rosbag record, send it a signal to stop, and check to see if
-  it cleaned up as expected.
-
-  :param test_bag_file_name: the output bag file
-  :param signal_to_test: the signal to send to rosbag
-  :return:
-  """
-  test_command = list(RECORD_COMMAND)
-  test_command.insert(4, test_bag_file_name)
-
-  p = subprocess.Popen(test_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
-  # wait while the recorder creates a bag for us to examine
-  time.sleep(SLEEP_TIME_SEC)
-  p.send_signal(signal_to_test)
-  p.wait()
+    self.assertTrue(os.path.isfile(TEST_BAG_FILE_NAME))
+    self.assertFalse(os.path.isfile(TEST_BAG_FILE_NAME+ '.active'))
 
 
 if __name__ == '__main__':
-  rostest.unitrun('test_rosbag', 'test_sigint_cleanup', RecordCleanup, sys.argv)
-  rostest.unitrun('test_rosbag', 'test_sigterm_cleanup', RecordCleanup, sys.argv)
+  rostest.unitrun('test_rosbag', 'test_sigterm_cleanup', RecordSigtermCleanup, sys.argv)
