+++ test/test_rosbag/test/record_sigterm_cleanup.test
@@ -1,6 +1,5 @@
 <launch>
   <node pkg="rostopic" type="rostopic" name="rostopic_pub1"
         args="pub -r 10 chatter std_msgs/String chatter1"/>
-  <test test-name="test_sigterm_cleanup" pkg="test_rosbag" type="record_cleanup.py"/>
+  <test test-name="test_sigterm_cleanup" pkg="test_rosbag" type="record_sigterm_cleanup.py"/>
 </launch>
-
